Fix time class API's and microseconds calculation.

+++ models/simconfig/include/SimpleSimConfig.hh
@@ -53,8 +53,8 @@ class SimpleSimConfig : public TrickHLA::Packing
    friend void init_attrTrickHLAModel__SimpleSimConfig();
 
   public:
-   double    run_duration;          ///< @trick_units{s}  The run duration of the simulation.
-   long long run_duration_microsec; ///< @trick_units{us} The run duration in microseconds.
+   double  run_duration;          ///< @trick_units{s}  The run duration of the simulation.
+   int64_t run_duration_microsec; ///< @trick_units{us} The run duration in microseconds.
 
    int   num_federates;      ///< @trick_units{--} Number of required federates.
    char *required_federates; ///< @trick_units{--} Comma-separated list of required federates.
